CodeClimate and i18n normalization issues fixed
